src/Docker/requirements.txt
requests
jsons
numpy
opencv-python
mediapipe
matplotlib
python-dotenv
Flask
flask-cors
zmq

